requirements.txt
fastapi
uvicorn
sse-starlette
pywebview
pyinstaller
Pillow
